Add knot to Ugly Tie shape
Added geometry to the Ugly Tie polygon to look like a knot. 

It remains a single polygon so color will affect both visible parts. 

At large zooms/resolutions a connection between the right side of the knot and the main tie is visible because the points on right side of the knot are not perfectly in-line with the upper right corner of the tie where the two larger parts of the shape are visible. If this becomes an issue, doing some math to find evenly dividing, aligned points near the current values would make the connecting section of the polygon zero width.